# internal/apidocs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Administrative REST API for managing the MCP Data Platform.\nEndpoints cover system info, configuration, personas, auth keys, audit logs, and knowledge management.",
        "title": "MCP Data Platform Admin API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1/admin",
    "paths": {
...
        "/config": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the current configuration as JSON with sensitive values redacted.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "summary": "Get config",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/admin.problemDetail"
                        }
                    }
                }
            }
        },
        "/config/export": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the current configuration as downloadable YAML. Sensitive values are redacted by default.",
                "produces": [
                    "application/x-yaml"
                ],
                "tags": [
                    "Config"
                ],
                "summary": "Export config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Set to 'true' to include sensitive values",
                        "name": "secrets",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/admin.problemDetail"
                        }
                    }
                }
            }
        },
        "/config/mode": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    },
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the current config store mode (file or database).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Config"
                ],
                "summary": "Get config mode",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/admin.configModeResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.configModeResponse": {
            "type": "object",
            "properties": {
                "mode": {
                    "type": "string",
                    "example": "database"
                },
                "read_only": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "admin.problemDetail": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "resource not found"
                },
                "status": {
                    "type": "integer",
                    "example": 404
                },
                "title": {
                    "type": "string",
                    "example": "Not Found"
                },
                "type": {
                    "type": "string",
                    "example": "about:blank"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}